fix(secrets): preserve Windows plan path trust
