+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -115,7 +115,7 @@ b6b8edc50ecab8386c9acd8f374a207212b5a99c8f518538bbcf0c458dda3881  module/runtime
 0d8f2c5f3a3325d7d190d2c395835f58ac57f1bea9b6fa516c6942b96d9bf605  module/runtime-store
 44adc2205f926172fcd3762ca8a96c1485beabcb1bef8b9acfd2233cefea2a6a  module/secret-input
 57dcb1462d4c4f9a98d934c4ca975b163d704758af9821a64001ff3ac05637c3  module/secret-input-runtime
-e249c8c5c7ce727b158c0cff1fff5515e7880788aa7d386b0ce58e8dbdf596d0  module/secret-ref-runtime
+e576b537880f63b3a91f3608f7e84c873bce6c6a3d9a0ba98c247f46de788d25  module/secret-ref-runtime
 596a315d426121c9620b314e3a9a7f523840b46e007d94d0d5e83cdedf789d15  module/security-runtime
 31b785e74f1f8f56241b7756ef6a5d86199c5ce177cbb1c234a261866972f270  module/session-discussion
 32fb6d253abf22440bc76c7a68d1f35fc0ef369b0ad738aedba9c3054a76e48e  module/session-store-runtime
